Added casts to reduce signed vs unsigned comparison warnings
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1970 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Network/network_microphone_shared.cpp
@@ -323,7 +323,7 @@ copy_and_send_audio_data(uint8* inFirstChunkReadPosition, int32 inFirstChunkByte
     int32 theTotalCaptureBytesConsumed = 0;
 
     // Keep sending if we have data and either we're squeezing out the last drop or we have a packet's-worth.
-    while(inFirstChunkBytesRemaining >= sCaptureBytesPerNetworkAudioByte &&
+    while(inFirstChunkBytesRemaining >= static_cast<int32>(sCaptureBytesPerNetworkAudioByte) &&
         (inForceSend || inFirstChunkBytesRemaining + inSecondChunkBytesRemaining >= (int32)sCaptureBytesPerPacket)) {
 
         theBytesConsumed = copy_data_in_capture_format_to_network_format(theOutgoingAudioData,
@@ -359,7 +359,7 @@ copy_and_send_audio_data(uint8* inFirstChunkReadPosition, int32 inFirstChunkByte
     }
 
     // Now, the first chunk is exhausted.  See if there's any left in the second chunk.  Same rules apply.
-    while(inSecondChunkBytesRemaining >= sCaptureBytesPerNetworkAudioByte &&
+    while(inSecondChunkBytesRemaining >= static_cast<int32>(sCaptureBytesPerNetworkAudioByte) &&
         (inForceSend || inSecondChunkBytesRemaining >= kNetworkAudioDataBytesPerPacket)) {
         
         theBytesConsumed = copy_data_in_capture_format_to_network_format(theOutgoingAudioData, kNetworkAudioDataBytesPerPacket,
